pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "oqd-cloud"
version = "0.1.0"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache 2.0"}
keywords = ["quantum", "computing", "simulation", "cloud", "server"]
classifiers = [
    "Development Status :: 3 - Alpha",

    "Intended Audience :: Science/Research ",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Information Analysis",

    "License :: OSI Approved :: Apache Software License",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]


dependencies = [
    "requests",
    "pydantic>=2.4",
    "oqd-core@git+https://github.com/OpenQuantumDesign/oqd-core",
    #
    "asyncpg",
    "uvicorn",
    "python-jose",
    "passlib",
    "python-multipart",
    "pydantic>=2.4",
    "sqlalchemy",
    "fastapi",
    "redis",
    "rq",
    "python-dotenv",
    "minio",
    #
    "oqd-analog-emulator@git+https://github.com/OpenQuantumDesign/oqd-analog-emulator",
    "oqd-trical@git+https://github.com/OpenQuantumDesign/oqd-trical",
]

[project.optional-dependencies]
docs = [
    "pymdown-extensions",
    "mkdocstrings",
    "mkdocs-material",
    "mkdocstrings-python",
    "mdx_truly_sane_lists"
]

tests = ["pytest"]

server = [
    "asyncpg",
    "uvicorn",
    "python-jose",
    "passlib",
    "python-multipart",
    "pydantic>=2.4",
    "sqlalchemy",
    "fastapi",
    "redis",
    "rq",
    "python-dotenv",
    "minio",
    #
    "oqd-analog-emulator@git+https://github.com/OpenQuantumDesign/oqd-analog-emulator",
    "oqd-trical@git+https://github.com/OpenQuantumDesign/oqd-trical",
]


[tool.setuptools.packages.find]
where = ["src"]
include = ["oqd_cloud*"]


[project.urls]
Homepage = "https://github.com/OpenQuantumDesign/oqd-cloud"
Repository = "https://github.com/OpenQuantumDesign/oqd-cloud.git"
Issues = "https://github.com/OpenQuantumDesign/oqd-cloud/issues"
